requirements.txt
python-aqi==0.6.1
git+https://github.com/ikalchev/py-sds011
